@@ -4,16 +4,16 @@ version: 0.0.4
 homepage: https://github.com/yoonjaepark/flutter_kakao_map
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: ^1.0.5
-  flutter_plugin_android_lifecycle: ^1.0.7
-  plugin_platform_interface: ^1.0.1
-  stream_transform: ^1.2.0
+  meta: ^1.7.0
+  flutter_plugin_android_lifecycle: ^2.0.3
+  plugin_platform_interface: ^2.0.1
+  stream_transform: ^2.0.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
