@@ -16,10 +16,10 @@ class KakaoMap extends StatefulWidget {
   ///
   /// [AssertionError] will be thrown if [initialCameraPosition] is null;
   const KakaoMap({
-    Key key,
-    @required this.initialCameraPosition,
+    Key? key,
+    required this.initialCameraPosition,
     this.onMapCreated,
-    this.gestureRecognizers,
+    required this.gestureRecognizers,
     this.compassEnabled = true,
     this.cameraTargetBounds = CameraTargetBounds.unbounded,
     this.mapType = MapType.standard,
@@ -39,21 +39,21 @@ class KakaoMap extends StatefulWidget {
     /// If no padding is specified default padding will be 0.
     this.padding = const EdgeInsets.all(0),
     this.indoorViewEnabled = false,
-    this.markers,
+    required this.markers,
     this.onCameraMoveStarted,
     this.onCameraMove,
     this.onCurrentLocationUpdate,
-    this.onCameraIdle,
-    this.onMarkerSelect,
-    this.onTap,
-    this.onLongPress,
+    required this.onCameraIdle,
+    required this.onMarkerSelect,
+    required this.onTap,
+    required this.onLongPress,
   })  : assert(initialCameraPosition != null),
         super(key: key);
 
   /// Callback method for when the map is ready to be used.
   ///
   /// Used to receive a [KakaoMapController] for this [KakaoMap].
-  final MapCreatedCallback onMapCreated;
+  final MapCreatedCallback? onMapCreated;
 
   /// 지도 카메라의 초기 위치입니다.
   final CameraPosition initialCameraPosition;
@@ -110,14 +110,14 @@ class KakaoMap extends StatefulWidget {
   /// 2. Programmatically initiated animation.
   /// 3. Camera motion initiated in response to user gestures on the map.
   ///    For example: pan, tilt, pinch to zoom, or rotate.
-  final VoidCallback onCameraMoveStarted;
+  final VoidCallback? onCameraMoveStarted;
 
   /// 지도 중심 좌표가 이동한 경우 호출된다.
-  final CameraPositionCallback onCameraMove;
+  final CameraPositionCallback? onCameraMove;
 
   /// CurrentLocationEventListener interface를 구현하는 객체를 MapView 객체에 등록하여
   /// 현위치 트래킹 이벤트를 통보받을 수 있다.
-  final CameraPositionCallback onCurrentLocationUpdate;
+  final CameraPositionCallback? onCurrentLocationUpdate;
 
   //단말 사용자가 POI Item을 선택한 경우 호출된다.
   // 사용자가 MapView 에 등록된 POI Item 아이콘(마커)를 터치한 경우 호출된다.
@@ -196,12 +196,12 @@ class _KakaoMapState extends State<KakaoMap> {
       Completer<KakaoMapController>();
 
   Map<MarkerId, Marker> _markers = <MarkerId, Marker>{};
-  _KakaoMapOptions _kakaoMapOptions;
+  late _KakaoMapOptions _kakaoMapOptions;
 
   @override
   Widget build(BuildContext context) {
     final Map<String, dynamic> creationParams = <String, dynamic>{
-      'initialCameraPosition': widget.initialCameraPosition?.toMap(),
+      'initialCameraPosition': widget.initialCameraPosition.toMap(),
       'options': _kakaoMapOptions.toMap(),
       'markersToAdd': serializeMarkerSet(widget.markers),
     };
@@ -255,43 +255,34 @@ class _KakaoMapState extends State<KakaoMap> {
     );
     _controller.complete(controller);
     if (widget.onMapCreated != null) {
-      widget.onMapCreated(controller);
+      widget.onMapCreated!(controller);
     }
   }
 
   void onMarkerTap(MarkerId markerId) {
-    assert(markerId != null);
     if (_markers[markerId]?.onTap != null) {
-      _markers[markerId].onTap();
+      _markers[markerId]!.onTap();
     }
   }
 
   void onMarkerDragEnd(MarkerId markerId, MapPoint position) {
-    assert(markerId != null);
     if (_markers[markerId]?.onDragEnd != null) {
-      _markers[markerId].onDragEnd(position);
+      _markers[markerId]!.onDragEnd(position);
     }
   }
 
   void onInfoWindowTap(MarkerId markerId) {
-    assert(markerId != null);
-    if (_markers[markerId]?.infoWindow?.onTap != null) {
-      _markers[markerId].infoWindow.onTap();
+    if (_markers[markerId]?.infoWindow.onTap != null) {
+      _markers[markerId]!.infoWindow.onTap!();
     }
   }
 
   void onTap(MapPoint position) {
-    assert(position != null);
-    if (widget.onTap != null) {
-      widget.onTap(position);
-    }
+    widget.onTap(position);
   }
 
   void onLongPress(MapPoint position) {
-    assert(position != null);
-    if (widget.onLongPress != null) {
-      widget.onLongPress(position);
-    }
+    widget.onLongPress(position);
   }
 }
 
@@ -301,22 +292,22 @@ class _KakaoMapState extends State<KakaoMap> {
 /// "do not change this configuration option".
 class _KakaoMapOptions {
   _KakaoMapOptions(
-      {this.compassEnabled,
-      this.cameraTargetBounds,
-      this.mapType,
-      this.currentLocationTrackingMode,
-      this.hdMapTileEnabled,
-      this.minMaxZoomPreference,
-      this.rotateGesturesEnabled,
-      this.scrollGesturesEnabled,
-      this.tiltGesturesEnabled,
-      this.trackCameraPosition,
-      this.zoomControlsEnabled,
-      this.zoomGesturesEnabled,
-      this.myLocationEnabled,
-      this.myLocationButtonEnabled,
-      this.padding,
-      this.indoorViewEnabled});
+      {required this.compassEnabled,
+      required this.cameraTargetBounds,
+      required this.mapType,
+      required this.currentLocationTrackingMode,
+      required this.hdMapTileEnabled,
+      required this.minMaxZoomPreference,
+      required this.rotateGesturesEnabled,
+      required this.scrollGesturesEnabled,
+      required this.tiltGesturesEnabled,
+      required this.trackCameraPosition,
+      required this.zoomControlsEnabled,
+      required this.zoomGesturesEnabled,
+      required this.myLocationEnabled,
+      required this.myLocationButtonEnabled,
+      required this.padding,
+      required this.indoorViewEnabled});
 
   static _KakaoMapOptions fromWidget(KakaoMap map) {
     return _KakaoMapOptions(
@@ -381,12 +372,12 @@ class _KakaoMapOptions {
     }
 
     addIfNonNull('compassEnabled', compassEnabled);
-    addIfNonNull('cameraTargetBounds', cameraTargetBounds?.toJson());
-    addIfNonNull('mapType', mapType?.index);
+    addIfNonNull('cameraTargetBounds', cameraTargetBounds.toJson());
+    addIfNonNull('mapType', mapType.index);
     addIfNonNull(
-        'currentLocationTrackingMode', currentLocationTrackingMode?.index);
+        'currentLocationTrackingMode', currentLocationTrackingMode.index);
     addIfNonNull('hdMapTile', hdMapTileEnabled);
-    addIfNonNull('minMaxZoomPreference', minMaxZoomPreference?.toJson());
+    addIfNonNull('minMaxZoomPreference', minMaxZoomPreference.toJson());
     addIfNonNull('rotateGesturesEnabled', rotateGesturesEnabled);
     addIfNonNull('scrollGesturesEnabled', scrollGesturesEnabled);
     addIfNonNull('tiltGesturesEnabled', tiltGesturesEnabled);
@@ -396,10 +387,10 @@ class _KakaoMapOptions {
     addIfNonNull('myLocationEnabled', myLocationEnabled);
     addIfNonNull('myLocationButtonEnabled', myLocationButtonEnabled);
     addIfNonNull('padding', <double>[
-      padding?.top,
-      padding?.left,
-      padding?.bottom,
-      padding?.right,
+      padding.top,
+      padding.left,
+      padding.bottom,
+      padding.right,
     ]);
     addIfNonNull('indoorEnabled', indoorViewEnabled);
     return optionsMap;
